# apps/api/internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	pgvector_go "github.com/pgvector/pgvector-go"
)

type ApiKey struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Name       string
	KeyHash    []byte
	Prefix     string
	CreatedAt  pgtype.Timestamptz
	LastUsedAt pgtype.Timestamptz
	RevokedAt  pgtype.Timestamptz
}

type Asset struct {
	ID               uuid.UUID
	UserID           uuid.UUID
	ItemID           pgtype.UUID
	ContentType      string
	ByteSize         int64
	OriginalFilename string
	CreatedAt        pgtype.Timestamptz
}

type DeviceLink struct {
	CodeHash   []byte
	UserID     uuid.UUID
	DeviceHint string
	CreatedAt  pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	ClaimedAt  pgtype.Timestamptz
}

type Feed struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	Url          string
	Title        string
	SiteUrl      string
	LastPolledAt pgtype.Timestamptz
	LastStatus   string
	CreatedAt    pgtype.Timestamptz
	Etag         string
	LastModified string
}

type Item struct {
	ID            uuid.UUID
	UserID        uuid.UUID
	Url           string
	Title         string
	Body          string
	LeadImageUrl  string
	Summary       string
	Tags          []string
	CardType      string
	Status        string
	CreatedAt     pgtype.Timestamptz
	UpdatedAt     pgtype.Timestamptz
	Palette       []string
	UserTags      []string
	PinnedAt      pgtype.Timestamptz
	LastDriftedAt pgtype.Timestamptz
	SearchTsv     interface{}
	PageCount     pgtype.Int4
}

type ItemEmbedding struct {
	ItemID    uuid.UUID
	UserID    uuid.UUID
	Embedding pgvector_go.Vector
	CreatedAt pgtype.Timestamptz
}

type Lense struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Name      string
	Rule      []byte
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type Link struct {
	UserID    uuid.UUID
	AItem     uuid.UUID
	BItem     uuid.UUID
	CreatedAt pgtype.Timestamptz
}

type User struct {
	ID          uuid.UUID
	CreatedAt   pgtype.Timestamptz
	ClerkUserID pgtype.Text
	Email       string
}