# openapi.yaml
openapi: 3.0.3
info: { title: Openmind API, version: 0.0.1 }
security:
  - bearerAuth: []
paths:
...
  /items:
    post:
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItemRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
    get:
      operationId: listItems
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Item" }
  /items/{id}:
    get:
      operationId: getItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "404":
          description: not found
    patch:
      operationId: patchItem
      description: "Edit an item. Currently supports replacing the full user-tags list (userTags). Tags are canonicalised server-side (trimmed, lowercased, deduped, capped)."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateItemRequest" }
      responses:
        "200":
          description: updated item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "400": { description: bad request }
        "404": { description: not found }
    delete:
      operationId: deleteItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: deleted }
        "404": { description: not found }
...
  /assets:
    post:
      operationId: createAsset
      description: "Upload a local image as a first-class image card (multipart field 'file')."
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: created image item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "400": { description: bad request }
        "413": { description: payload too large }
        "415": { description: unsupported media type }
  /assets/{id}:
    get:
      operationId: getAsset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: image bytes
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
            image/gif: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
            image/avif: { schema: { type: string, format: binary } }
        "404": { description: not found }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    CreateItemRequest:
      type: object
      description: "Exactly one of url or note must be provided."
      properties:
        url: { type: string, format: uri }
        note: { type: string, minLength: 1, maxLength: 10000 }
    Item:
      type: object
      required: [id, url, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        title: { type: string }
        summary: { type: string }
        leadImageUrl: { type: string }
        tags: { type: array, items: { type: string } }
        userTags: { type: array, items: { type: string } }
        palette: { type: array, items: { type: string } }
        cardType: { type: string, enum: [article, product, book, recipe, video, tweet, image, note, quote] }
        status: { type: string, enum: [pending, enriched, failed] }
        createdAt: { type: string, format: date-time }
        pinnedAt: { type: string, format: date-time, nullable: true, description: "When the item was pinned to the Desk; null if not pinned." }
        pageCount: { type: integer, nullable: true, description: "Page count when the item's original is a stored PDF; null otherwise." }
    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/Item"
        - type: object
          required: [body]
          properties:
            body: { type: string }
    UpdateItemRequest:
      type: object
      description: "Fields to update on an item. Both userTags and pinned are optional; omit both for a no-op edit that is rejected as a bad request."
      properties:
        userTags: { type: array, items: { type: string }, description: "Full replacement user-tags list. An empty array clears all user tags." }
        pinned: { type: boolean, description: "Pin (true) or unpin (false) the item on the Desk. Pinning sets pinnedAt to now; unpinning clears it." }
...